数据库Long类型修改为Integer类型

+++ xiaohe66-web-portal/src/main/java/com/xiaohe66/web/base/base/BaseDao.java
@@ -28,7 +28,7 @@ public interface BaseDao<T extends BasePo>{
      * 通用根据id删除方法(硬删除，数据库不留数据)
      * @param id 待删除的数据库id
      */
-    void delByIdOfHard(long id);
+    void delByIdOfHard(Integer id);
 
     /**
      * 通用根据id删除方法(软删除，数据库保留数据)
@@ -36,14 +36,14 @@ public interface BaseDao<T extends BasePo>{
      * @param currentUsrId 当前登录用户id
      * @param updateTime 删除时间
      */
-    void delById(@Param("id") long id, @Param("currentUsrId") long currentUsrId, @Param("updateTime") Date updateTime);
+    void delById(@Param("id") Integer id, @Param("currentUsrId") Integer currentUsrId, @Param("updateTime") Date updateTime);
 
     /**
      * 通用根据参数删除方法(硬删除，数据库不留数据)
      * @param param 传入mybatis的参数
-     * @return long 删除的数量
+     * @return Integer 删除的数量
      */
-    long delByParamOfHard(@Param("param") BaseParam param);
+    Integer delByParamOfHard(@Param("param") BaseParam param);
 
     /**
      * 通用根据参数删除方法(软删除，数据库保留数据)
@@ -51,9 +51,9 @@ public interface BaseDao<T extends BasePo>{
      * @param currentUsrId 当前登录用户id
      * @param updateTime 删除时间
      *
-     * @return long 删除的数量
+     * @return Integer 删除的数量
      */
-    long delByParam(@Param("param") BaseParam param, @Param("currentUsrId") long currentUsrId, @Param("updateTime") Date updateTime);
+    Integer delByParam(@Param("param") BaseParam param, @Param("currentUsrId") Integer currentUsrId, @Param("updateTime") Date updateTime);
 
     /**
      * 通用更新方法
@@ -73,7 +73,7 @@ public interface BaseDao<T extends BasePo>{
      * @param id 查询id
      * @return 对应实体
      */
-    T findById(Long id);
+    T findById(Integer id);
 
     /**
      * 通用根据参数查询方法
@@ -86,5 +86,5 @@ public interface BaseDao<T extends BasePo>{
      * 数据统计
      * @return  返回表的数据数量
      */
-    Long count();
+    Integer count();
 }
