+++ xiaohe66-web-portal/src/main/java/com/xiaohe66/web/base/base/BaseService.java
@@ -14,49 +14,49 @@ public interface BaseService<T extends BasePo> {
      * @param po 插入的实体
      * @param currentUsrId 当前操作者id
      */
-    void add(T po, Long currentUsrId);
+    void add(T po, Integer currentUsrId);
 
     /**
      * 通用批量插入方法
      * @param list 批量插入的实体list
      * @param currentUsrId 当前操作者id
      */
-    void addAll(List<T> list, Long currentUsrId);
+    void addAll(List<T> list, Integer currentUsrId);
 
     /**
      * 通用根据id删除方法(硬删除，数据库不留数据)
      * @param id 待删除的数据库id
      */
-    void delByIdOfHard(Long id);
+    void delByIdOfHard(Integer id);
 
     /**
      * 通用根据id删除方法(软删除，数据库保留数据)
      * @param id 待删除的数据库id
      * @param currentUsrId 当前登录用户id
      */
-    void delById(Long id, Long currentUsrId);
+    void delById(Integer id, Integer currentUsrId);
 
     /*
      * 通用根据参数删除方法(硬删除，数据库不留数据)
      * @param param 传入mybatis的参数
-     * @return long 删除的数量
+     * @return Integer 删除的数量
      */
-//    long delByParamOfHard(BaseParam param);
+//    Integer delByParamOfHard(BaseParam param);
 
     /*
      * 通用根据参数删除方法(软删除，数据库保留数据)
      * @param param 传入mybatis的参数
      * @param currentUsrId 当前登录用户id
-     * @return long 删除的数量
+     * @return Integer 删除的数量
      */
-//    long delByParam(BaseParam param,Long currentUsrId);
+//    Integer delByParam(BaseParam param,Integer currentUsrId);
 
     /**
      * 通用更新方法
      * @param po 更新后的实体
      * @param currentUsrId 当前操作者id
      */
-    void updateById(T po, Long currentUsrId);
+    void updateById(T po, Integer currentUsrId);
 
     /*
      * 通用更新方法，根据参数更新
@@ -64,14 +64,14 @@ public interface BaseService<T extends BasePo> {
      * @param param 更新的参数
      * @param currentUsrId 当前操作者id
      */
-//    void updateByParam(T po, BaseParam param, Long currentUsrId);
+//    void updateByParam(T po, BaseParam param, Integer currentUsrId);
 
     /**
      * 通用根据id查询方法
      * @param id 查询id
      * @return 对应实体
      */
-    T findById(Long id);
+    T findById(Integer id);
 
     /*
      * 通用根据参数查询方法
@@ -84,5 +84,5 @@ public interface BaseService<T extends BasePo> {
      * 数据统计
      * @return  返回表的数据数量
      */
-    Long count();
+    Integer count();
 }
